folder

+++ app/Http/Controllers/Upload/UploadController.php
@@ -38,8 +38,35 @@
         /*
          *删除文件
          * */
-        public function delete(){
-
+        public function delete(Request $request){
+            if(Session::has('id')){
+                $uid = Session::get('id');
+                $fileid= $request->input('fileid')? $request->input('fileid') : "";
+                if($uid && $fileid){
+                    $sql01 = "SELECT * FROM files WHERE  fileid=? and uid=?";
+                    $res01 =  DB::select($sql01,[$fileid,$uid]);
+                    if($res01){
+                        $result  = @unlink($res01[0]->filepath);
+                        if(!$result){
+                            $sql02 = "DELETE FROM files WHERE fileid=? and uid=?";
+                            $res02 = DB::delete($sql02,[$fileid,$uid]);
+                            if ($res02){
+                                return response("1");
+                            }else{
+                                return response("-5");
+                            }
+                        }else{
+                            return response("-4");
+                        }
+                    }else{
+                        return response("-3");
+                    }
+                }else{
+                    return response("-2");
+                }
+            }else{
+                return response("-1");
+            }
         }
 
         /*
@@ -90,18 +117,50 @@
         /*
          * 删除文件夹
          * */
-        public function deletefolder(Request $request){
+        /**
+         * @param Request $request
+         * @return \Illuminate\Contracts\Routing\ResponseFactory|\Symfony\Component\HttpFoundation\Response
+         */
+        public function deletefolder( Request $request){
             $uid = Session::get('id');
             $foldername = $request->input('foldername')? $request->input('foldername') : "";
             $foldergrade = $request->input('foldergrade')? $request->input('foldergrade') : "";
             $folderid = $request->input('folderid')? $request->input('folderid') : "";
-            if($uid && $foldername && $foldergrade ){
-                $sql = "DELETE FROM folders where ";
+            if($uid && $foldername && $foldergrade && $folderid ){
+                /*递归查找
+                  * */
+              /*  $arr = array();
+                function digui($id, $arr,$index,$jijie){
+                    $sql =  "select folid from folders where folpreid='$id'";
+                    $res = DB::select($sql);
+                    if($res) {
+                        for($i = 0;$i < count($res);$i++){
+                            $arr[$index]['id'] = $res[$i]->folid;
+                            $arr[$index]['jijie'] = $jijie;
+                            $index++;
+                            digui($res[$i]->folid,$arr,$index,$jijie+1);
+                        }
+                    }else{
+                        return $arr;
+                    }
+                }*/
+                $sql01 = "select folid from folders where folpreid = ? and uid=? ";
+                $res01 = DB::select($sql01,[$folderid.$uid]);
+                if(count($res01) >= 2){
+                    return response("-2");
+                }else{
+                    $sql = "DELETE FROM folders where folid=? and grade=? and uid=?";
+                    $res = DB::delete($sql,[$folderid,$foldergrade,$uid]);
+                    if($res){
+                        return response("1");
+                    }else{
+                        return response("-3");
+                    }
+                }
             }else{
                 return response("-1");
             }
         }
-
         /*
          * 保存到数据库里面
          * */
